update payment logic and enrollment information.

+++ controllers/paymentRecordsController.js
@@ -3,6 +3,8 @@ import AppError from './../utils/appError.js';
 import APIFeatures from './../utils/apiKeyFeatures.js';
 import PaymentRecord from '../models/paymentRecordsMode.js';
 import Wishlist from '../models/wishlistModel.js';
+import Enroll from '../models/courseEnrollModel.js';
+import Course from '../models/coursesModel.js';
 
 export const getAllPaymentRecords = catchAsync(async (req, res, next) => {
   const features = new APIFeatures(PaymentRecord.find(), req.query)
@@ -35,6 +37,11 @@ export const getPaymentRecord = catchAsync(async (req, res, next) => {
 });
 
 export const createPaymentRecord = catchAsync(async (req, res, next) => {
+  //check if the course exist
+  const course = await Course.findById(req.body.course);
+  if (!course) {
+    return next(new AppError('No course found with that ID', 404));
+  }
   const newPaymentRecord = await PaymentRecord.create({
     user: req.user._id,
     course: req.body.course,
@@ -50,6 +57,15 @@ export const createPaymentRecord = catchAsync(async (req, res, next) => {
       new AppError('Invalid data  please provide all required data', 400),
     );
   }
+  const previousEnrollments = await Enroll.findOne({
+    course,
+    user: req.user._id,
+  });
+  if (previousEnrollments) {
+    return next(new AppError('User already enroll on this course', 400));
+  }
+
+  await Enroll.create({ course: req.body.course, user: req.user._id });
   // after creating the payment record check if the user has this course in his wishlist and remove it
   await Wishlist.findOneAndDelete({
     user: req.user._id,
